requirements.txt
Django>=3.2.1,<3.3.0
djangorestframework>=3.13.0,<3.14.0
markdown>=3.3.6,<3.4.0
psycopg2>=2.9.3,<3.0.0
PyYAML>=6.0.0,<6.1.0

